create separated JobAdContent component
