+++ src/@next/pages/CreateJobAd/index.tsx
@@ -1,11 +1,3 @@
-import {
-  Box,
-  Container,
-  Grid,
-  makeStyles,
-  Theme,
-  Typography,
-} from "@material-ui/core";
 import { useAuth, useCart, useCheckout } from "@saleor/sdk";
 import { useRouter } from "next/router";
 import React from "react";
@@ -15,23 +7,9 @@ import {
   apolloClient,
   createCheckoutQuery,
 } from "@components/organisms/SetTargetGroup/queries";
-
-import JobCriteriaForm from "./JobCriteriaForm";
-import JobPostingDetailsForm from "./JobPostingDetailsForm";
-
-const useStyles = makeStyles<Theme>(theme => ({
-  root: {
-    display: "flex",
-    flexWrap: "wrap",
-  },
-  header: {
-    marginTop: theme.spacing(5),
-    marginBottom: theme.spacing(2),
-  },
-}));
+import CreateJobAdContent from "@pages/CreateJobAd/CreateJobAdContent";
 
 const CreateJobAd = () => {
-  const classes = useStyles();
   const { replace } = useRouter();
   const { user } = useAuth();
   const { loaded: cartLoaded, items } = useCart();
@@ -86,25 +64,7 @@ const CreateJobAd = () => {
     return null;
   }
 
-  return (
-    <Container maxWidth="xl" className={classes.root}>
-      <Grid container justify="center">
-        <Grid item xs={8}>
-          <Box className={classes.header}>
-            <Typography variant="h4" component="h1">
-              Create your job ad(s)
-            </Typography>
-            <Typography variant="body1" component="h6">
-              Fill in the needed information for your job posting(s). Some
-              fields are required by job boards to be able to post on it.
-            </Typography>
-          </Box>
-          <JobPostingDetailsForm />
-          <JobCriteriaForm />
-        </Grid>
-      </Grid>
-    </Container>
-  );
+  return <CreateJobAdContent />;
 };
 
 export default CreateJobAd;
